Ruby Objects Diagnostic
Clarity 2
Comfort 1

I'm hitting a really flat wall with Ruby and I really want to spend some one on one
or group review time trying to understand even the tiniest bit more of what we've
gone over in the past week. Some of js stuck bits at a time and then found cohesion
through time and repetition but I feel like not even the smallest bits of Ruby are
sticking with me yet.